+++ lib/l10n/app_localizations_fr.dart
@@ -826,4 +826,7 @@ class AppLocalizationsFr extends AppLocalizations {
 
   @override
   String get mustChangePassword => 'Must change password on first login';
+
+  @override
+  String get pleaseEnterTitle => 'Please enter a title';
 }